requirements.txt
Django==4.2.6
requests==2.31.0
djangorestframework==3.14.0
paho-mqtt==1.6.1